# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/responsible/me/face": {
      "get": {
        "operationId": "FaceEnrollmentController_getMyFace",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Status da face do responsável (foto cadastrada e sync com leitores)",
        "tags": [
          "responsible-face-enrollment"
        ]
      },
      "post": {
        "operationId": "FaceEnrollmentController_uploadMyFace",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enviar/atualizar foto do responsável e sincronizar com os leitores",
        "tags": [
          "responsible-face-enrollment"
        ]
      }
    },
    "/responsible/me/face/sync": {
      "post": {
        "operationId": "FaceEnrollmentController_resyncMyFace",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reenviar foto já armazenada para os leitores faciais (sem nova imagem)",
        "tags": [
          "responsible-face-enrollment"
        ]
      }
    },
    "/responsible/children/{studentId}/face": {
      "get": {
        "operationId": "FaceEnrollmentController_getChildFace",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Status da face de um aluno vinculado ao responsável",
        "tags": [
          "responsible-face-enrollment"
        ]
      },
      "post": {
        "operationId": "FaceEnrollmentController_uploadChildFace",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enviar/atualizar foto de um aluno vinculado e sincronizar com os leitores",
        "tags": [
          "responsible-face-enrollment"
        ]
      }
    },
    "/responsible/children/{studentId}/face/sync": {
      "post": {
        "operationId": "FaceEnrollmentController_resyncChildFace",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reenviar foto do aluno já armazenada para os leitores (sem nova imagem)",
        "tags": [
          "responsible-face-enrollment"
        ]
      }
    },
    "/responsible/household/{responsibleId}/face": {
      "get": {
        "operationId": "FaceEnrollmentController_getHouseholdMemberFace",
        "parameters": [
          {
            "name": "responsibleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Status da face de um co-responsável do mesmo núcleo (compartilham alunos)",
        "tags": [
          "responsible-face-enrollment"
        ]
      },
      "post": {
        "operationId": "FaceEnrollmentController_uploadHouseholdMemberFace",
        "parameters": [
          {
            "name": "responsibleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enviar/atualizar foto de um co-responsável do núcleo e sincronizar com os leitores",
        "tags": [
          "responsible-face-enrollment"
        ]
      }
    },
    "/responsible/household/{responsibleId}/face/sync": {
      "post": {
        "operationId": "FaceEnrollmentController_resyncHouseholdMemberFace",
        "parameters": [
          {
            "name": "responsibleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reenviar foto do co-responsável já armazenada para os leitores (sem nova imagem)",
        "tags": [
          "responsible-face-enrollment"
        ]
      }
    },
    "/member/me/face": {
      "get": {
        "operationId": "MemberFaceEnrollmentController_getMyFace",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Status da face do membro (foto cadastrada e sync com leitores)",
        "tags": [
          "member-face-enrollment"
        ]
      },
      "post": {
        "operationId": "MemberFaceEnrollmentController_uploadMyFace",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enviar/atualizar foto do membro e sincronizar com os leitores",
        "tags": [
          "member-face-enrollment"
        ]
      }
    },
    "/member/me/face/sync": {
      "post": {
        "operationId": "MemberFaceEnrollmentController_resyncMyFace",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reenviar foto já armazenada para os leitores faciais (sem nova imagem)",
        "tags": [
          "member-face-enrollment"
        ]
      }
    },
...
  },
  "info": {
    "title": "Face2go API",
    "description": "Backend NestJS da plataforma Face2go",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UploadFaceDto": {
        "type": "object",
        "properties": {
          "imageBase64": {
            "type": "string",
            "minLength": 64,
            "description": "Imagem JPEG em base64 (opcional prefixo data:image/jpeg;base64,)"
          }
        },
        "required": [
          "imageBase64"
        ]
      },
...
    }
  }
}